Move `RemoveIssueModal` to its own directory
This component now has a few accompanying files and I feel is
justified to put it in its own directory.
